Add pop-up functionality to User Profile
In this commit i added a pop-up functionality to <ProfileView /> component.
The reactjs-popup library was used in achieving this,
When a user clicks on his/her username from the <ProfileView /> component, a form is poped up.
This poped-up form (<EditProfile /> component) is used to edit/update the details of that specific user